[Ibex] Disable Bitmanip in Libraries
We now build the toolchain (including libraries) with only rv32imc (as
before we landed the bitmanip patches). With the B extension compiler
and binutils patches landed, the compiler should still be able to
compile for cores with the +B extension, but this change prevents issues
if your core does not implement the B extension, which is the case for
mainline Ibex today.

Code with the B extension enabled should be linkable with libraries with
the B extension disabled.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>