Downgrade RV32IMC GCC to 9.2.0
With the previous config, we were accidentally building a GCC 10.0.0
experimental version of GCC. In order to make the toolchain migration go
as smoothly as possible for OpenTitan, I'm reverting to building the
same version as we were building before (9.2.0).

I will follow-up with a patch that includes the config for RV32IMC+B, as
this does require a git version of GCC because we apply patches on top
of it.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>