Check Tarballs for Broken Symlinks
We've had problems with releasing tarballs with broken symlinks in them,
which the prior commits have aimed to solve. This commit adds a
regression test to ensure the build fails when there are broken symlinks
in a release tarball.

We do this after uploading azure artifacts (which are not easy to find),
but before uploading to github releases, so if a test fails we can find
the tarball to debug, but the tarball is not released to the public.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>