Add script to run build locally
The CI builds of the toolchain are done by Azure Pipelines in a
container, which is specially prepared by Azure Pipelines.  The build
steps run as normal user (not root) which has sudo permissions. We have
prepared our base container to provide similar functionality for local
use cases; running the build through this script can be used to get a
local development experience similar to the Azure Pipelines runner.