[crosstool-ng] Use upstream ct-ng version with ISL mirror fixes
This ct-ng version (the tip of master, at the time of this commit) includes
a recent ct-ng commit (cfb7d07a) that fixes the mirror URL of ISL. Without
it ISL would fail to download. Centos 6 build machines still require a
separate fix for expired Let's Encrypt CA certificates that is out of the
scope of this patch.

With this update the upstream ct-ng suffices to build the our GNU
toolchains so we no longer need to use a fork. The update requires the
following minimal version changes:

- GCC: 9.2.0 -> 9.4.0
- GDB: 8.2.1 -> 8.3.1