Move tarballing step out of GCC/Clang build scripts
The output directory is reused in CI so that the Clang tarball includes
the GCC and Binutils parts of the toolchain. It doesn't make sense to me
to prepare the tarballs inside the build scripts when they bleed into
each other.

This commit changes CI to prepare a single tarball including GCC, Clang,
and Binutils.