Update Binutils to 2.44
Upstream binutils now supports the ratified 1.0.0 bitmanip extensions,
so we now use upstream and add the unratified extensions in the patch.

The updated binutils also contain a fix for generating verilog memory
(vmem) files that is useful for OpenTitan.

The glob and ELF BFD patches are no longer required for this version of
binutils to build correctly so they are removed.

Signed-off-by: Alice Ziuziakowska <a.ziuziakowska@lowrisc.org>