Update Binutils to 2.44
Upstream binutils now supports the ratified 1.0.0 bitmanip extensions,
so rather than maintaining patches to add the unratified extensions back
in, a fork of binutils was created which we now use, and should allow
for easier updates to future versions.

The glob and ELF BFD patches are no longer required for this version of
binutils to build correctly so they are removed.

Since the last version of binutils used, the GMP and MPFR libraries have
become required for building GDB. However, the configure step will
always prefer the shared library rather than the static one when
building, so we have to build our own static-only libraries of these to
link.

MPFR is clone-able from the official repository, however GMP does not
appreciate excessive cloning, so we use a release archive instead for
that.

The updated binutils contain a change for generating verilog memory
(vmem) files that is useful for OpenTitan.

See also: https://github.com/lowRISC/binutils

Signed-off-by: Alice Ziuziakowska <a.ziuziakowska@lowrisc.org>